First draft of the new version lc2.py
Maybe that "first draft" is a bit of a simplification. The current state
is still not functional, hence the draft. But there is enough code to
justify adding the file to the repository.

The old files could not work with the last versions of Inkscape, due to
changes in the inkex libraries. This version is independent from Inkscape,
and produces a SVG file that can be debugged by text or with tools like
inkview.

And it is supposed to be more configurable, and more documented (the old
code generating the SVG was frankly a mess).