Compute the inside holes of the plates
This commit looks good, but is still not completely
functional not thoroughly tested. It computes the
holes inside the plates, for a normally well-fitting
of the plates.

There is still a special case that is not handled :
if two interior plates meet on both sides of another
interior plate. In this case, the crenelations will
at the moment overlap.